Initial commit in 0.6 branch.
Permission model is removed. Roles' permissions are stored in a serialized attribute of the Role model.
There are now 2 builtin roles: *Non member* and *Anonymous*. They are used to specify permissions given to other users than members of projects.

For example, it's now possible:
* to restrict wiki edition to loggedin users only, by removing 'Edit wiki' permission from 'Anonymous' role.
* to restrict wiki edition to project members only, by removing 'Edit wiki' permission from 'Non member' and 'Anonymous' roles.

Permissions have to be re-set after applying this change (go to Admin -> Roles & permissions -> Permissions report).

git-svn-id: https://svn.redmine.org/redmine@639 e93f8b46-1217-0410-a6f0-8f06a7374b81